ENT-12880: Created new integration test targets to enable builds to go faster (#5364)
* ENT-12880: Created new target integrationTest1 & integrationTest2 to enable builds to go faster.
