+++ node/src/integration-test1/kotlin/net/corda/node/services/vault/VaultObserverExceptionTest.kt
@@ -781,7 +781,7 @@ class VaultObserverExceptionTest {
         ) {
             val aliceNode = startNode(providedName = ALICE_NAME, rpcUsers = listOf(user)).getOrThrow()
 
-            val flowHandle = aliceNode.rpc.startFlow(::SubscribingRawUpdatesFlow)
+            val flowHandle = aliceNode.rpc.startFlow(VaultObserverExceptionTest::SubscribingRawUpdatesFlow)
 
             assertFailsWith<CordaRuntimeException>(
                 "Flow ${SubscribingRawUpdatesFlow::class.java.name} tried to access VaultService.rawUpdates " +
