+++ node/src/integration-test2/kotlin/net/corda/contracts/serialization/missing/CustomData.kt
