+++ node/src/integration-test2/kotlin/net/corda/flows/serialization/missing/CustomDataSerializer.kt
