+++ node/src/integration-test2/kotlin/net/corda/node/ContractWithMissingCustomSerializerTest.kt
@@ -18,6 +18,7 @@ import net.corda.testing.core.DUMMY_NOTARY_NAME
 import net.corda.testing.driver.DriverParameters
 import net.corda.testing.driver.driver
 import net.corda.testing.driver.internal.incrementalPortAllocation
+import net.corda.testing.internal.assertNotCordaSerializable
 import net.corda.testing.node.NotarySpec
 import net.corda.testing.node.TestCordapp
 import net.corda.testing.node.User
