+++ node/src/integration-test2/kotlin/net/corda/node/services/statemachine/FlowHospitalTest.kt
